chore: move onboarding components to dedicated folder for better project structure
