+++ src/app/(auth)/sign-up/(onboarding)/components/OnboardingContainer.tsx
@@ -5,9 +5,9 @@ import { useState } from "react";
 import { ArrowLeftIcon, ArrowRightIcon } from "@heroicons/react/24/outline";
 import Image from "next/image";
 import WelcomeChingu from "./WelcomeChingu";
-import WhatFeatures from "./WhatFeatures";
-import HowDidYouHear from "./HowDidYouHear";
-import LinkedInUrl from "./LinkedInUrl";
+import WhatFeatures from "@/app/(auth)/sign-up/(onboarding)/components/WhatFeatures";
+import HowDidYouHear from "@/app/(auth)/sign-up/(onboarding)/components/HowDidYouHear";
+import LinkedInUrl from "@/app/(auth)/sign-up/(onboarding)/components/LinkedInUrl";
 import Button from "@/components/Button";
 import type { SteppersItem } from "@/components/Stepper";
 import Stepper from "@/components/Stepper";
