+++ src/app/(auth)/sign-up/(onboarding)/components/WelcomeChingu.tsx
