+++ src/app/(auth)/sign-up/(onboarding)/components/WhatFeatures.tsx
